a function to remove the temp markers
